split out files, add recusion detection test
