+++ print3rs-commands/src/commands/version.rs
@@ -1,0 +1,4 @@
+pub fn version() -> &'static str {
+    const VERSION: Option<&str> = option_env!("CARGO_PKG_VERSION");
+    VERSION.unwrap_or("???")
+}
