+++ print3rs-commands/src/lib.rs
@@ -1,3 +1,1 @@
-
 pub mod commands;
-pub mod logging;
