--- print3rs-commands/src/logging/mod.rs
@@ -1,4 +1,0 @@
-pub mod parsing;
-
-
-//pub fn start_logging(name: &str, pattern_parts: &[Segment]) ->
